verify admin before delete operation

+++ index.js
@@ -16,6 +16,26 @@ app.use(express.json());
 const uri = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASS}@cluster0.e3n1sso.mongodb.net/?retryWrites=true&w=majority`;
 const client = new MongoClient(uri, { useNewUrlParser: true, useUnifiedTopology: true, serverApi: ServerApiVersion.v1 });
 
+// jwt function/middleware
+function verifyJWT(req, res, next) {
+    // console.log(req.headers.authorization);
+
+    const authHeader = req.headers.authorization;
+    if (!authHeader) {
+        return res.status(401).send('unauthorized access');
+    }
+
+    const token = authHeader.split(' ')[1];
+
+    jwt.verify(token, process.env.ACCESS_TOKEN, function (err, decoded) {
+        if (err) {
+            return res.status(403).send({ message: 'forbidden access' })
+        }
+        req.decoded = decoded;
+        next();
+    })
+}
+
 async function run() {
     try {
         //  -------------------------------- collections ---------------------------------
@@ -68,8 +88,16 @@ async function run() {
         })
 
         // get my bookings 
-        app.get('/bookings', async (req, res) => {
+        app.get('/bookings', verifyJWT, async (req, res) => {
             const email = req.query.email;
+            // console.log(req.headers.authorization);
+
+            const decodedEmail = req.decoded.email;
+
+            if (email !== decodedEmail) {
+                return res.status(403).send({ message: 'forbidden access' })
+            }
+
             const query = { email: email }
             const result = await bookingsCollection.find(query).toArray()
             res.send(result);
@@ -179,7 +207,17 @@ async function run() {
         })
 
         // delete buyer
-        app.delete('/users/buyers/:id', async (req, res) => {
+        app.delete('/users/buyers/:id', verifyJWT, async (req, res) => {
+
+            // first checking if the role is "admin"
+            const decodedEmail = req.decoded.email;
+            const query = { email: decodedEmail };
+            const user = await usersCollection.findOne(query);
+
+            if (user?.role !== 'Admin') {
+                return res.status(403).send({ message: 'forbidden access' })
+            }
+
             const id = req.params.id;
             const filter = { _id: ObjectId(id) };
             const result = await usersCollection.deleteOne(filter);
@@ -196,7 +234,18 @@ async function run() {
         })
 
         // delete seller
-        app.delete('/users/sellers/:id', async (req, res) => {
+        app.delete('/users/sellers/:id', verifyJWT, async (req, res) => {
+
+            // first checking if the role is "admin"
+            const decodedEmail = req.decoded.email;
+            const query = { email: decodedEmail };
+            const user = await usersCollection.findOne(query);
+
+            if (user?.role !== 'Admin') {
+                return res.status(403).send({ message: 'forbidden access' })
+            }
+
+            // now if the user is "admin" then delete will be applied
             const id = req.params.id;
             const filter = { _id: ObjectId(id) };
             const result = await usersCollection.deleteOne(filter);
